GPG-709 Moved ReportingYearsHelper to Core library
