+++ GenderPayGap.WebUI/Classes/BaseClasses/BaseController.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using System.Web;
 using GenderPayGap.Core;
 using GenderPayGap.Core.Classes.Logger;
+using GenderPayGap.Core.Helpers;
 using GenderPayGap.Core.Interfaces;
 using GenderPayGap.Core.Models;
 using GenderPayGap.Core.Models.HttpResultModels;
