+++ GenderPayGap.WebUI/Controllers/Report/ReportDiscardDraftController.cs
@@ -1,4 +1,4 @@
-﻿using GenderPayGap.Core;
+﻿using GenderPayGap.Core.Helpers;
 using GenderPayGap.Core.Interfaces;
 using GenderPayGap.Database;
 using GenderPayGap.Database.Models;
