+++ GenderPayGap.WebUI/Controllers/Report/ReportOverviewController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using GenderPayGap.Core;
+using GenderPayGap.Core.Helpers;
 using GenderPayGap.Core.Interfaces;
 using GenderPayGap.Database;
 using GenderPayGap.Database.Models;
