+++ GenderPayGap.WebUI/Controllers/Report/ReportResponsiblePersonController.cs
@@ -1,4 +1,5 @@
 ﻿using GenderPayGap.Core;
+using GenderPayGap.Core.Helpers;
 using GenderPayGap.Core.Interfaces;
 using GenderPayGap.Database;
 using GenderPayGap.Database.Models;
