+++ GenderPayGap.WebUI/Controllers/Report/ReportReviewAndSubmitController.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-using GenderPayGap.Core;
 using GenderPayGap.Core.Classes;
+using GenderPayGap.Core.Helpers;
 using GenderPayGap.Core.Interfaces;
 using GenderPayGap.Database;
 using GenderPayGap.Database.Models;
