+++ GenderPayGap.WebUI/Helpers/ControllerHelper.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Security.Claims;
 using GenderPayGap.Core;
+using GenderPayGap.Core.Helpers;
 using GenderPayGap.Core.Interfaces;
 using GenderPayGap.Database;
 using GenderPayGap.WebUI.Classes;
