+++ GenderPayGap.WebUI/Views/AccountCreation/ConfirmEmailAddress.cshtml
@@ -1,4 +1,4 @@
-﻿@using GenderPayGap.Core
+﻿@using GenderPayGap.Core.Helpers
 @model GenderPayGap.WebUI.Models.AccountCreation.ConfirmEmailAddressViewModel
 @{
     ViewBag.Title = "Confirm your email address - Gender pay gap service";
