+++ GenderPayGap.WebUI/Views/AccountCreation/ConfirmationPage.cshtml
@@ -1,4 +1,4 @@
-@using GenderPayGap.Core
+@using GenderPayGap.Core.Helpers
 @model GenderPayGap.WebUI.Models.AccountCreation.ConfirmEmailAddressViewModel
 @{
     ViewBag.Title = "You've confirmed your email address - Gender pay gap service";
