+++ GenderPayGap.WebUI/Views/AccountCreation/CreateUserAccount.cshtml
@@ -1,7 +1,7 @@
-@using GenderPayGap.Core
-@using GenderPayGap.WebUI.Models.AccountCreation
 @using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
+@using GenderPayGap.Core.Helpers
+@using GenderPayGap.WebUI.Models.AccountCreation
 @model GenderPayGap.WebUI.Models.AccountCreation.CreateUserAccountViewModel
 @{
     ViewBag.Title = "Create my account - Gender pay gap reporting service";
