+++ GenderPayGap.WebUI/Views/AccountCreation/UserNotFoundErrorPage.cshtml
@@ -1,4 +1,4 @@
-﻿@using GenderPayGap.Core
+﻿@using GenderPayGap.Core.Helpers
 @{
     ViewBag.Title = "User not found - Gender pay gap reporting service";
     Layout = "~/Views/GovUkFrontend/GovUkFrontendLayout.cshtml";
