+++ GenderPayGap.WebUI/Views/ActionHub/Sections/RelatedContent.cshtml
@@ -1,5 +1,5 @@
-﻿@using GenderPayGap.Core
-
+﻿@using GenderPayGap.Core.Helpers
+@using GenderPayGap.Core
 <div class="guidance-related-content">
     <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible related-content-section-break">
