+++ GenderPayGap.WebUI/Views/Admin/ViewOrganisation.cshtml
@@ -1,9 +1,9 @@
-@using GenderPayGap.Core
-@using GenderPayGap.Database
-@using GenderPayGap.WebUI.Helpers
+@using Microsoft.AspNetCore.Http
 @using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
-@using Microsoft.AspNetCore.Http
+@using GenderPayGap.Core
+@using GenderPayGap.Core.Helpers
+@using GenderPayGap.Database
 @model GenderPayGap.WebUI.Models.Admin.AdminViewOrganisationViewModel
 
 @{
