+++ GenderPayGap.WebUI/Views/AdminOrganisationScope/ChangeScope.cshtml
@@ -1,8 +1,8 @@
-@using GenderPayGap.Core
-@using GenderPayGap.WebUI.Helpers
-@using GenderPayGap.WebUI.Models.Admin
 @using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
+@using GenderPayGap.WebUI.Models.Admin
+@using GenderPayGap.Core.Helpers
+@using GenderPayGap.Core
 @model GenderPayGap.WebUI.Models.Admin.AdminChangeScopeViewModel
 
 @{
