+++ GenderPayGap.WebUI/Views/AdminOrganisationScope/ViewOrganisationScope.cshtml
@@ -1,7 +1,7 @@
-﻿@using GenderPayGap.Database
-@using GenderPayGap.WebUI.Helpers
-@using GovUkDesignSystem
+﻿@using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
+@using GenderPayGap.Database
+@using GenderPayGap.Core.Helpers
 @model GenderPayGap.Database.Organisation
 
 @{
