+++ GenderPayGap.WebUI/Views/GOVUK/Layouts/template.cshtml
@@ -1,7 +1,7 @@
-@using GenderPayGap.Core
-@using GenderPayGap.WebUI.Controllers.Login
-@using GenderPayGap.WebUI.Cookies
 @using GenderPayGap.WebUI.Helpers
+@using GenderPayGap.WebUI.Cookies
+@using GenderPayGap.Core.Helpers
+@using GenderPayGap.WebUI.Controllers.Login
 @{
     string assetsPath = "/assets/images";
