+++ GenderPayGap.WebUI/Views/GovUkFrontend/GovUkFrontendLayout.cshtml
@@ -1,8 +1,8 @@
-﻿@using GenderPayGap.Core
-@using GenderPayGap.WebUI.Cookies
-@using GenderPayGap.WebUI.Helpers
-@using GovUkDesignSystem
+﻿@using GovUkDesignSystem
 @using GovUkDesignSystem.GovUkDesignSystemComponents
+@using GenderPayGap.WebUI.Helpers
+@using GenderPayGap.WebUI.Cookies
+@using GenderPayGap.Core.Helpers
 @{
     string assetsPath = "/assets/images";
 }
