+++ GenderPayGap.WebUI/Views/ManageOrganisations/OrganisationReportsTable.cshtml
@@ -1,5 +1,5 @@
-﻿@using GenderPayGap.Core
-@using GenderPayGap.WebUI.Models.ManageOrganisations
+﻿@using GenderPayGap.WebUI.Models.ManageOrganisations
+@using GenderPayGap.Core.Helpers
 @model GenderPayGap.WebUI.Models.ManageOrganisations.ManageOrganisationViewModel
 
 <table class="govuk-table">
