+++ GenderPayGap.WebUI/Views/Organisation/DisplayScopeStatus.cshtml
@@ -1,5 +1,5 @@
 ﻿@using GenderPayGap.Core
-@using GenderPayGap.WebUI.Helpers
+@using GenderPayGap.Core.Helpers
 @model dynamic
 @{
     ScopeStatuses scopeStatus = Model.scopeStatus;
