+++ GenderPayGap.WebUI/Views/Organisation/ManageOrganisation.cshtml
@@ -1,7 +1,7 @@
-@using GenderPayGap.Core
 @using GenderPayGap.Database
-@using GenderPayGap.WebUI.Helpers
 @using GenderPayGap.WebUI.Models.Organisation
+@using GenderPayGap.Core
+@using GenderPayGap.Core.Helpers
 @model GenderPayGap.WebUI.Models.Organisation.ManageOrganisationModel
 @{
     var controller = ViewData["Controller"] as OrganisationController;
