+++ GenderPayGap.WebUI/Views/Viewing/Download.cshtml
@@ -1,5 +1,4 @@
-﻿@using GenderPayGap.WebUI.Helpers
-
+﻿@using GenderPayGap.Core.Helpers
 @{
     ViewBag.Title = "Download gender pay gap figures - Gender pay gap service";
     Layout = "~/Views/GovUkFrontend/GovUkFrontendLayout.cshtml";
