+++ GenderPayGap.WebUI/Views/Viewing/EmployerDetails/Employer.cshtml
@@ -1,7 +1,7 @@
-﻿@using GenderPayGap.Core
-@using GenderPayGap.Database
-@using GenderPayGap.WebUI.Helpers
+﻿@using GenderPayGap.Database
 @using GenderPayGap.WebUI.Models.Shared
+@using GenderPayGap.Core
+@using GenderPayGap.Core.Helpers
 @model EmployerDetailsViewModel
 @{
     Organisation organisation = Model.Organisation;
